+++ internal/cli/arg_context.go
@@ -1,5 +1,7 @@
 package cli
 
+import "fmt"
+
 
 // A custom integer type for an Enum to keep track of the arguments context.
 type argContext int
@@ -14,14 +16,18 @@ const (
 
   // The context where arguments are interpreted as mapping files.
   contextMapFile
+
+  // The context where arguments are interpreted as a mapping.
+  contextMapping
 )
 
 // Get an argContext as a human readable string.
 func (context argContext) String() string {
   names := []string{
     "Unknown",
-    "--config/-c",
-    "--map/-m",
+    fmt.Sprintf("%s/%s", configOption, configOptionAlias),
+    fmt.Sprintf("%s/%s", mapfileOption, mapfileOptionAlias),
+    fmt.Sprintf("%s/%s", mappingOption, mappingOptionAlias),
   }
 
   return names[context]