+++ internal/cli/parse_test.go
@@ -49,6 +49,36 @@ func testDefaultMapFiles(t *testing.T, arguments Arguments) {
   }
 }
 
+func testDefaultMappings(t *testing.T, arguments Arguments) {
+  if len(arguments.Mappings) != 0 {
+    t.Error("The default list of Mappings should be empty")
+  }
+}
+
+func testDefaultsExcept(t *testing.T, arguments Arguments, exclude string) {
+  if exclude != "dryrun" {
+    testDefaultDryRun(t, arguments)
+  }
+  if exclude != "invert" {
+    testDefaultInvert(t, arguments)
+  }
+  if exclude != "silent" {
+    testDefaultSilent(t, arguments)
+  }
+  if exclude != "verbose" {
+    testDefaultVerbose(t, arguments)
+  }
+  if exclude != "config file" {
+    testDefaultConfigFile(t, arguments)
+  }
+  if exclude != "map files" {
+    testDefaultMapFiles(t, arguments)
+  }
+  if exclude != "mappings" {
+    testDefaultMappings(t, arguments)
+  }
+}
+
 
 func ExampleParseArgs() {
   args := []string {"wordrow", "foo.bar"}
@@ -68,7 +98,7 @@ func TestNoArgs(t *testing.T) {
 }
 
 func TestHelpArg(t *testing.T) {
-  args := createArgs(helpOption)
+  args := createArgs(helpFlag)
   run, _ := ParseArgs(args)
 
   if run == true {
@@ -78,7 +108,7 @@ func TestHelpArg(t *testing.T) {
 
 func TestVersionArg(t *testing.T) {
   t.Run("--version only", func(t *testing.T) {
-    args := createArgs(versionOption)
+    args := createArgs(versionFlag)
     run, _ := ParseArgs(args)
 
     if run == true {
@@ -86,7 +116,7 @@ func TestVersionArg(t *testing.T) {
     }
   })
   t.Run("--version and other", func(t *testing.T) {
-    args := createArgs(versionOption, "foo.bar")
+    args := createArgs(versionFlag, "foo.bar")
     run, _ := ParseArgs(args)
 
     if run != true {
@@ -103,12 +133,7 @@ func TestDefaultOptions(t *testing.T) {
     t.Fatal("The first return value should be true for this test")
   }
 
-  testDefaultDryRun(t, arguments)
-  testDefaultInvert(t, arguments)
-  testDefaultSilent(t, arguments)
-  testDefaultVerbose(t, arguments)
-  testDefaultConfigFile(t, arguments)
-  testDefaultMapFiles(t, arguments)
+  testDefaultsExcept(t, arguments, "")
 
   if len(arguments.InputFiles) != 1 {
     t.Error("The list of InputFiles should contain a single file")
@@ -116,142 +141,114 @@ func TestDefaultOptions(t *testing.T) {
 }
 
 func TestDryRunFlag(t *testing.T) {
-  args := createArgs(dryRunOption, "foo.bar")
+  args := createArgs(dryRunFlag, "foo.bar")
   run, arguments := ParseArgs(args)
 
   if run != true {
     t.Fatal("The first return value should be true for this test")
   }
 
-  testDefaultInvert(t, arguments)
-  testDefaultSilent(t, arguments)
-  testDefaultVerbose(t, arguments)
-  testDefaultConfigFile(t, arguments)
-  testDefaultMapFiles(t, arguments)
+  testDefaultsExcept(t, arguments, "dryrun")
 
   if arguments.DryRun != true {
-    t.Errorf("The DryRun value should be true if %s is an argument", dryRunOption)
+    t.Errorf("The DryRun value should be true if %s is an argument", dryRunFlag)
   }
 }
 
 func TestInvertFlag(t *testing.T) {
-  t.Run(invertOption, func(t *testing.T) {
-    args := createArgs(invertOption, "foo.bar")
+  t.Run(invertFlag, func(t *testing.T) {
+    args := createArgs(invertFlag, "foo.bar")
     run, arguments := ParseArgs(args)
 
     if run != true {
       t.Fatal("The first return value should be true for this test")
     }
 
-    testDefaultDryRun(t, arguments)
-    testDefaultSilent(t, arguments)
-    testDefaultVerbose(t, arguments)
-    testDefaultConfigFile(t, arguments)
-    testDefaultMapFiles(t, arguments)
+    testDefaultsExcept(t, arguments, "invert")
 
     if arguments.Invert != true {
-      t.Errorf("The Invert value should be true if %s is an argument", invertOption)
+      t.Errorf("The Invert value should be true if %s is an argument", invertFlag)
     }
   })
-  t.Run(invertOptionAlias, func(t *testing.T) {
-    args := createArgs(invertOptionAlias, "foo.bar")
+  t.Run(invertFlagAlias, func(t *testing.T) {
+    args := createArgs(invertFlagAlias, "foo.bar")
     run, arguments := ParseArgs(args)
 
     if run != true {
       t.Fatal("The first return value should be true for this test")
     }
 
-    testDefaultDryRun(t, arguments)
-    testDefaultSilent(t, arguments)
-    testDefaultVerbose(t, arguments)
-    testDefaultConfigFile(t, arguments)
-    testDefaultMapFiles(t, arguments)
+    testDefaultsExcept(t, arguments, "invert")
 
     if arguments.Invert != true {
-      t.Errorf("The Invert value should be true if %s is an argument", invertOptionAlias)
+      t.Errorf("The Invert value should be true if %s is an argument", invertFlagAlias)
     }
   })
 }
 
 func TestSilentFlag(t *testing.T) {
-  t.Run(silentOption, func(t *testing.T) {
-    args := createArgs(silentOption, "foo.bar")
+  t.Run(silentFlag, func(t *testing.T) {
+    args := createArgs(silentFlag, "foo.bar")
     run, arguments := ParseArgs(args)
 
     if run != true {
       t.Fatal("The first return value should be true for this test")
     }
 
-    testDefaultDryRun(t, arguments)
-    testDefaultInvert(t, arguments)
-    testDefaultVerbose(t, arguments)
-    testDefaultConfigFile(t, arguments)
-    testDefaultMapFiles(t, arguments)
+    testDefaultsExcept(t, arguments, "silent")
 
     if arguments.Silent != true {
-      t.Errorf("The Silent value should be true if %s is an argument", silentOption)
+      t.Errorf("The Silent value should be true if %s is an argument", silentFlag)
     }
   })
-  t.Run(silentOptionAlias, func(t *testing.T) {
-    args := createArgs(silentOptionAlias, "foo.bar")
+  t.Run(silentFlagAlias, func(t *testing.T) {
+    args := createArgs(silentFlagAlias, "foo.bar")
     run, arguments := ParseArgs(args)
 
     if run != true {
       t.Fatal("The first return value should be true for this test")
     }
 
-    testDefaultDryRun(t, arguments)
-    testDefaultInvert(t, arguments)
-    testDefaultVerbose(t, arguments)
-    testDefaultConfigFile(t, arguments)
-    testDefaultMapFiles(t, arguments)
+    testDefaultsExcept(t, arguments, "silent")
 
     if arguments.Silent != true {
-      t.Errorf("The Silent value should be true if %s is an argument", silentOptionAlias)
+      t.Errorf("The Silent value should be true if %s is an argument", silentFlagAlias)
     }
   })
 }
 
 func TestVerboseFlag(t *testing.T) {
-  t.Run(verboseOption, func(t *testing.T) {
-    args := createArgs(verboseOption, "foo.bar")
+  t.Run(verboseFlag, func(t *testing.T) {
+    args := createArgs(verboseFlag, "foo.bar")
     run, arguments := ParseArgs(args)
 
     if run != true {
       t.Fatal("The first return value should be true for this test")
     }
 
-    testDefaultDryRun(t, arguments)
-    testDefaultInvert(t, arguments)
-    testDefaultSilent(t, arguments)
-    testDefaultConfigFile(t, arguments)
-    testDefaultMapFiles(t, arguments)
+    testDefaultsExcept(t, arguments, "verbose")
 
     if arguments.Verbose != true {
-      t.Errorf("The Verbose value should be true if %s is an argument", verboseOption)
+      t.Errorf("The Verbose value should be true if %s is an argument", verboseFlag)
     }
   })
-  t.Run(verboseOptionAlias, func(t *testing.T) {
-    args := createArgs(verboseOptionAlias, "foo.bar")
+  t.Run(verboseFlagAlias, func(t *testing.T) {
+    args := createArgs(verboseFlagAlias, "foo.bar")
     run, arguments := ParseArgs(args)
 
     if run != true {
       t.Fatal("The first return value should be true for this test")
     }
 
-    testDefaultDryRun(t, arguments)
-    testDefaultInvert(t, arguments)
-    testDefaultSilent(t, arguments)
-    testDefaultConfigFile(t, arguments)
-    testDefaultMapFiles(t, arguments)
+    testDefaultsExcept(t, arguments, "verbose")
 
     if arguments.Verbose != true {
-      t.Errorf("The Verbose value should be true if %s is an argument", verboseOptionAlias)
+      t.Errorf("The Verbose value should be true if %s is an argument", verboseFlagAlias)
     }
   })
 }
 
-func TestConfigOption(t *testing.T) {
+func TestConfigFileOption(t *testing.T) {
   configFile := "config.json"
 
   t.Run(configOption, func(t *testing.T) {
@@ -262,11 +259,7 @@ func TestConfigOption(t *testing.T) {
       t.Fatal("The first return value should be true for this test")
     }
 
-    testDefaultDryRun(t, arguments)
-    testDefaultInvert(t, arguments)
-    testDefaultSilent(t, arguments)
-    testDefaultVerbose(t, arguments)
-    testDefaultMapFiles(t, arguments)
+    testDefaultsExcept(t, arguments, "config file")
 
     if arguments.ConfigFile != configFile {
       t.Errorf("The config file was incorrect (was '%s')", arguments.ConfigFile)
@@ -280,11 +273,7 @@ func TestConfigOption(t *testing.T) {
       t.Fatal("The first return value should be true for this test")
     }
 
-    testDefaultDryRun(t, arguments)
-    testDefaultInvert(t, arguments)
-    testDefaultSilent(t, arguments)
-    testDefaultVerbose(t, arguments)
-    testDefaultMapFiles(t, arguments)
+    testDefaultsExcept(t, arguments, "config file")
 
     if arguments.ConfigFile != configFile {
       t.Errorf("The config file was incorrect (was '%s')", arguments.ConfigFile)
@@ -300,11 +289,7 @@ func TestConfigOption(t *testing.T) {
       t.Fatal("The first return value should be true for this test")
     }
 
-    testDefaultDryRun(t, arguments)
-    testDefaultInvert(t, arguments)
-    testDefaultSilent(t, arguments)
-    testDefaultVerbose(t, arguments)
-    testDefaultMapFiles(t, arguments)
+    testDefaultsExcept(t, arguments, "config file")
 
     if arguments.ConfigFile != otherConfigFile {
       t.Errorf("The config file was incorrect (was '%s')", arguments.ConfigFile)
@@ -312,7 +297,7 @@ func TestConfigOption(t *testing.T) {
   })
 }
 
-func TestConfigOptionIncorrect(t *testing.T) {
+func TestConfigFileOptionIncorrect(t *testing.T) {
   t.Run("value missing", func(t *testing.T) {
     args := createArgs(configOption)
     run, _ := ParseArgs(args)
@@ -322,7 +307,7 @@ func TestConfigOptionIncorrect(t *testing.T) {
     }
   })
   t.Run("other flag", func(t *testing.T) {
-    args := createArgs(configOption, silentOption)
+    args := createArgs(configOption, silentFlag)
     run, _ := ParseArgs(args)
 
     if run != false {
@@ -331,7 +316,7 @@ func TestConfigOptionIncorrect(t *testing.T) {
   })
 }
 
-func TestMappingOption(t *testing.T) {
+func TestMapFileOption(t *testing.T) {
   mapFile := "foo.map"
 
   t.Run(mapfileOption, func(t *testing.T) {
@@ -342,11 +327,7 @@ func TestMappingOption(t *testing.T) {
       t.Fatal("The first return value should be true for this test")
     }
 
-    testDefaultDryRun(t, arguments)
-    testDefaultInvert(t, arguments)
-    testDefaultSilent(t, arguments)
-    testDefaultVerbose(t, arguments)
-    testDefaultConfigFile(t, arguments)
+    testDefaultsExcept(t, arguments, "map files")
 
     if len(arguments.MapFiles) != 1 {
       t.Fatalf("The MapFiles list have length 1 (was %d)", len(arguments.MapFiles))
@@ -364,11 +345,7 @@ func TestMappingOption(t *testing.T) {
       t.Fatal("The first return value should be true for this test")
     }
 
-    testDefaultDryRun(t, arguments)
-    testDefaultInvert(t, arguments)
-    testDefaultSilent(t, arguments)
-    testDefaultVerbose(t, arguments)
-    testDefaultConfigFile(t, arguments)
+    testDefaultsExcept(t, arguments, "map files")
 
     if len(arguments.MapFiles) != 1 {
       t.Fatalf("The MapFiles list have length 1 (was %d)", len(arguments.MapFiles))
@@ -388,11 +365,7 @@ func TestMappingOption(t *testing.T) {
       t.Fatal("The first return value should be true for this test")
     }
 
-    testDefaultDryRun(t, arguments)
-    testDefaultInvert(t, arguments)
-    testDefaultSilent(t, arguments)
-    testDefaultVerbose(t, arguments)
-    testDefaultConfigFile(t, arguments)
+    testDefaultsExcept(t, arguments, "map files")
 
     if len(arguments.MapFiles) != 2 {
       t.Fatalf("The MapFiles list have length 2 (was %d)", len(arguments.MapFiles))
@@ -408,7 +381,7 @@ func TestMappingOption(t *testing.T) {
   })
 }
 
-func TestMappingOptionIncorrect(t *testing.T) {
+func TestMapFileOptionIncorrect(t *testing.T) {
   t.Run("value missing", func(t *testing.T) {
     args := createArgs(mapfileOption)
     run, _ := ParseArgs(args)
@@ -418,7 +391,91 @@ func TestMappingOptionIncorrect(t *testing.T) {
     }
   })
   t.Run("other flag", func(t *testing.T) {
-    args := createArgs(mapfileOption, silentOption)
+    args := createArgs(mapfileOption, silentFlag)
+    run, _ := ParseArgs(args)
+
+    if run != false {
+      t.Error("The first return value should be false if there is an error in the args")
+    }
+  })
+}
+
+func TestMappingOption(t *testing.T) {
+  mapping := "foo,bar"
+
+  t.Run(mappingOption, func(t *testing.T) {
+    args := createArgs(mappingOption, mapping, "foo.bar")
+    run, arguments := ParseArgs(args)
+
+    if run != true {
+      t.Fatal("The first return value should be true for this test")
+    }
+
+    testDefaultsExcept(t, arguments, "mappings")
+
+    if len(arguments.Mappings) != 1 {
+      t.Fatalf("The Mappings list have length 1 (was %d)", len(arguments.Mappings))
+    }
+
+    if arguments.Mappings[0] != mapping {
+      t.Errorf("First mapping was incorrect (was '%s')", arguments.Mappings[0])
+    }
+  })
+  t.Run(mappingOptionAlias, func(t *testing.T) {
+    args := createArgs(mappingOptionAlias, mapping, "foo.bar")
+    run, arguments := ParseArgs(args)
+
+    if run != true {
+      t.Fatal("The first return value should be true for this test")
+    }
+
+    testDefaultsExcept(t, arguments, "mappings")
+
+    if len(arguments.Mappings) != 1 {
+      t.Fatalf("The Mappings list have length 1 (was %d)", len(arguments.Mappings))
+    }
+
+    if arguments.Mappings[0] != mapping {
+      t.Errorf("First mapping was incorrect (was '%s')", arguments.Mappings[0])
+    }
+  })
+  t.Run("multiple mappings", func(t *testing.T) {
+    otherMapping := "cat,dog"
+
+    args := createArgs(mappingOption, mapping, mappingOptionAlias, otherMapping, "foo.bar")
+    run, arguments := ParseArgs(args)
+
+    if run != true {
+      t.Fatal("The first return value should be true for this test")
+    }
+
+    testDefaultsExcept(t, arguments, "mappings")
+
+    if len(arguments.Mappings) != 2 {
+      t.Fatalf("The Mappings list have length 1 (was %d)", len(arguments.Mappings))
+    }
+
+    if arguments.Mappings[0] != mapping {
+      t.Errorf("First mapping was incorrect (was '%s')", arguments.Mappings[0])
+    }
+
+    if arguments.Mappings[1] != otherMapping {
+      t.Errorf("First mapping was incorrect (was '%s')", arguments.Mappings[0])
+    }
+  })
+}
+
+func TestMappingOptionIncorrect(t *testing.T) {
+  t.Run("value missing", func(t *testing.T) {
+    args := createArgs(mappingOption)
+    run, _ := ParseArgs(args)
+
+    if run != false {
+      t.Error("The first return value should be false if there is an error in the args")
+    }
+  })
+  t.Run("other flag", func(t *testing.T) {
+    args := createArgs(mappingOption, silentFlag)
     run, _ := ParseArgs(args)
 
     if run != false {