+++ internal/fs/paths.go
@@ -8,6 +8,8 @@ import "regexp"
 import "github.com/ericcornelissen/wordrow/internal/errors"
 import "github.com/ericcornelissen/wordrow/internal/logger"
 
+import "github.com/yargevad/filepathx"
+
 
 // Regular expression for glob strings.
 var globExpr = regexp.MustCompile("[\\*\\?\\[\\]]")
@@ -26,6 +28,11 @@ func getCwd() string {
   return cwd
 }
 
+// Get the extension of a file give the file path.
+func getExt(path string) string {
+  return filepath.Ext(path)
+}
+
 
 // Resolve any number of globs or file paths into distinct file paths.
 //
@@ -39,7 +46,7 @@ func ResolveGlobs(patterns ...string) (paths []string, rerr error) {
       continue
     }
 
-    matches, err := filepath.Glob(pattern)
+    matches, err := filepathx.Glob(pattern)
     if err != nil {
       rerr = errors.Newf("Malformed pattern (%s)", pattern)
     } else {
@@ -50,20 +57,28 @@ func ResolveGlobs(patterns ...string) (paths []string, rerr error) {
   return paths, rerr
 }
 
+// Resolve a single absolute or relative path to an absolute path.
+//
+// The function panics if the (current) working directory is needed but could
+// not be found.
+func ResolvePath(inputPath string) string {
+  if filepath.IsAbs(inputPath) {
+    return inputPath
+  } else {
+    outputPath := path.Join(getCwd(), inputPath)
+    return outputPath
+  }
+}
+
 // Resolve any number of absolute or relative paths to absolute paths only.
 //
 // The function panics if the (current) working directory is needed but could
 // not be found.
-func ResolvePaths(files ...string) []string {
+func ResolvePaths(inputPaths ...string) []string {
   var paths []string
-  for i := 0; i < len(files); i++ {
-    file := files[i]
-    if filepath.IsAbs(file) {
-      paths = append(paths, file)
-    } else {
-      file = path.Join(getCwd(), file)
-      paths = append(paths, file)
-    }
+  for _, inputPath := range inputPaths {
+    outputPath := ResolvePath(inputPath)
+    paths = append(paths, outputPath)
   }
 
   return paths